Simple tests (#11606)
* method to simplify migrate/create tests
* use new assertion for all migration/create tests
* move the expected files to test/data/console
