+++ tests/data/console/migrate_create/create_test.php
@@ -1,0 +1,32 @@
+<?php
+
+return <<<CODE
+<?php
+
+use yii\db\Migration;
+
+/**
+ * Handles the creation for table `test`.
+ */
+class {$class} extends Migration
+{
+    /**
+     * @inheritdoc
+     */
+    public function up()
+    {
+        \$this->createTable('test', [
+            'id' => \$this->primaryKey(),
+        ]);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function down()
+    {
+        \$this->dropTable('test');
+    }
+}
+
+CODE;
