+++ tests/data/console/migrate_create/create_title_pk.php
@@ -1,0 +1,34 @@
+<?php
+
+return <<<CODE
+<?php
+
+use yii\db\Migration;
+
+/**
+ * Handles the creation for table `test`.
+ */
+class {$class} extends Migration
+{
+    /**
+     * @inheritdoc
+     */
+    public function up()
+    {
+        \$this->createTable('test', [
+            'title' => \$this->primaryKey(),
+            'body' => \$this->text()->notNull(),
+            'price' => \$this->money(11,2),
+        ]);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function down()
+    {
+        \$this->dropTable('test');
+    }
+}
+
+CODE;
