+++ tests/data/console/migrate_create/drop_test.php
@@ -1,0 +1,32 @@
+<?php
+
+return <<<CODE
+<?php
+
+use yii\db\Migration;
+
+/**
+ * Handles the dropping for table `test`.
+ */
+class {$class} extends Migration
+{
+    /**
+     * @inheritdoc
+     */
+    public function up()
+    {
+        \$this->dropTable('test');
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function down()
+    {
+        \$this->createTable('test', [
+            'id' => \$this->primaryKey(),
+        ]);
+    }
+}
+
+CODE;
